move MXUtil class
