+++ runelite-client/src/main/java/net/runelite/client/plugins/microbot/maxxin/astralrc/AstralRunesScript.java
@@ -6,7 +6,7 @@ import net.runelite.api.coords.WorldPoint;
 import net.runelite.api.gameval.ItemID;
 import net.runelite.client.plugins.microbot.Microbot;
 import net.runelite.client.plugins.microbot.Script;
-import net.runelite.client.plugins.microbot.maxxin.util.MXUtil;
+import net.runelite.client.plugins.microbot.maxxin.MXUtil;
 import net.runelite.client.plugins.microbot.util.antiban.Rs2Antiban;
 import net.runelite.client.plugins.microbot.util.antiban.Rs2AntibanSettings;
 import net.runelite.client.plugins.microbot.util.bank.Rs2Bank;
